Add WalkToKickBall strategy

+++ module/purpose/Defender/src/Defender.cpp
@@ -107,7 +107,7 @@ namespace module::purpose {
             // If we are stable, walk to the ready field position
             log<NUClear::DEBUG>("READY");
             emit<Task>(std::make_unique<WalkToFieldPosition>(
-                Eigen::Vector3f(cfg.ready_position.x(), cfg.ready_position.y(), 0),
+                Eigen::Vector3d(cfg.ready_position.x(), cfg.ready_position.y(), 0),
                 cfg.ready_position.z()));
         });
